refactor: move SegmentationMerger interface to segment_dti_abnormal_values module and update references
